Fix browser bundling issue
